Improved YouTube Video Forced Resolution method
For some reason, setPlaybackQuality seems to be super finicky with the IFrame API and Awesomium, so now we're just going to "suggest" the resolution on loadVideoById, since it seems much more reliable. We're keeping the Think-based stuff as a way to update it in realtime still...if that's still possible.